services/qa-pairing/requirements.txt
torch==2.6.0 --index-url https://download.pytorch.org/whl/cu121
torchvision==0.21.0 --index-url https://download.pytorch.org/whl/cu121
torchaudio==2.6.0 --index-url https://download.pytorch.org/whl/cu121
transformers==4.46.3
tokenizers==0.20.3
flask>=3.0.0
flask-cors>=4.0.0
Pillow>=10.0.0
addict>=2.4.0
matplotlib>=3.7.0
einops>=0.7.0
easydict>=1.10
safetensors>=0.4.0
accelerate>=0.25.0
hf-xet>=0.1.0
pdfplumber>=0.10.0
pdf2image>=1.16.0
requests>=2.31.0
PyMuPDF>=1.23.0
python-dotenv>=1.0.0
